Ignore non-existent container in deleteContainer
All other blobstore providers allow this operation.  Further this
matches the behavior of TransientStorageStrategy.deleteContainer.
This commit prevents a spurious error message from
deleteContainerOrWarnIfUnable after test suite completion.

+++ apis/filesystem/src/main/java/org/jclouds/filesystem/strategy/internal/FilesystemStorageStrategyImpl.java
@@ -131,6 +131,9 @@ public class FilesystemStorageStrategyImpl implements LocalStorageStrategy {
    @Override
    public void deleteContainer(String container) {
       filesystemContainerNameValidator.validate(container);
+      if (!containerExists(container)) {
+         return;
+      }
       deleteDirectory(container, null);
    }
